+++ docs/changelog.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+
+## New in 2.9.0 (2025-01-15)
+
 - Update Action Server to `2.5.1`
 - Add `sema4ai.fixWrongAgentImport` command
 - Show Output button when there are errors while running action or agent package commands